requirements.txt
django
djangorestframework
social-auth-app-django
Pillow